update from origin

+++ app/exts/manager.js
@@ -7,12 +7,16 @@ import db from './extensions-db';
 import {createExtension} from './extension';
 
 const createSavePath = extension => {
-    return Path.join(env.dataPath, 'xexts', extension.name);
+    return extension.localPath || Path.join(env.dataPath, 'xexts', extension.name);
 };
 
 const uninstall = extension => {
-    const savedPath = createSavePath(extension);
     return db.removeInstall(extension).then(() => {
+        extension.detach();
+        if (extension.isDev) {
+            return Promise.resolve();
+        }
+        const savedPath = createSavePath(extension);
         return fse.remove(savedPath);
     });
 };
@@ -31,25 +35,54 @@ const extractInstallFile = filePath => {
     });
 };
 
-const installFromDir = (dir, deleteDir = false) => {
+const reloadDevExtension = extension => {
+    const path = extension.localPath;
+    if (path) {
+        const pkgFilePath = Path.join(path, 'package.json');
+        const pkg = fse.readJSONSync(pkgFilePath, {throws: false});
+        if (pkg) {
+            extension = createExtension(pkg, extension.data);
+            db.saveInstall(extension, true);
+            if (DEBUG) {
+                console.collapse('Extension Reload for Dev', 'greenBg', extension.name, 'greenPale');
+                console.log('extension', extension);
+                console.groupEnd();
+            }
+            return extension;
+        }
+    }
+    return false;
+};
+
+const installFromDir = (dir, deleteDir = false, devMode = false) => {
     const pkgFilePath = Path.join(dir, 'package.json');
+    const savedPath = devMode ? dir : createSavePath(extension);
     let extension = null;
     return fse.readJSON(pkgFilePath).then(pkg => {
-        extension = createExtension(pkg);
-        return db.saveInstall(extension);
-    }).then(extension => {
-        const savedPath = createSavePath(extension);
-        return fse.emptyDir(savedPath).then(() => {
-            return fse.copy(dir, savedPath);
+        extension = createExtension(pkg, {
+            localPath: savedPath,
+            isDev: devMode
         });
+        return db.saveInstall(extension);
+    }).then(() => {
+        if (!devMode) {
+            return fse.emptyDir(savedPath).then(() => {
+                return fse.copy(dir, savedPath);
+            });
+        }
+        return Promise.resolve(extension);
     }).then(() => {
         if (deleteDir) {
-            return fse.remove(dir);
+            return fse.remove(dir).then(() => {
+                return Promise.resolve(extension);
+            });
         }
-        return Promise.resolve();
-    }).then(() => {
         return Promise.resolve(extension);
     });
+};
+
+const installFromDevDir = (dir) => {
+    return installFromDir(dir, false, true);
 };
 
 const installFromXextFile = filePath => {
@@ -58,13 +91,13 @@ const installFromXextFile = filePath => {
     });
 };
 
-const openInstallDialog = (callback) => {
-    dialog.showOpenDialog('.xext,.json,.zip', files => {
+const openInstallDialog = (callback, devMode = false) => {
+    dialog.showOpenDialog(devMode ? '.json' : '.xext,.json,.zip', files => {
         if (files.length) {
             const filePath = files[0].path;
             const extName = Path.extname(filePath).toLowerCase();
             if (extName === '.json' && Path.basename(filePath) === 'package.json') {
-                installFromDir(Path.dirname(filePath)).then(extension => {
+                installFromDir(Path.dirname(filePath), false, devMode).then(extension => {
                     if (callback) callback(extension);
                 }).catch(error => {
                     if (callback) callback(false, error);
@@ -100,5 +133,7 @@ export default {
     installFromXextFile,
     openInstallDialog,
     loadReadmeMarkdown,
+    installFromDevDir,
+    reloadDevExtension,
 };
